+++ src/utils/colors.js
@@ -1,0 +1,5 @@
+export const COLORS = {
+  blue: "#1E5BC6",
+  grey: "#d3d3d3",
+  orange: "#ec8643",
+};