+++ CHANGELOG.md
@@ -16,3 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Card shadow when hovering.
+
+### Changed
+
+- Color of navbar/card/button/link